Use MyProgressBar
I just want to override onDraw. Cherry-picked from one that translated
offscreen. Original commit message below:

Use a translate call to push the AVD offscreen

Though the View is on screen, the AVD itself has a matrix. Verify that
we respect that matrix.

Custom MyProgressBar just does a translate that happens to be the width
of the screen of my test device.

Conflicts:
- app/src/main/res/layout/activity_main.xml