Cherry-pick of "Use a translate call to push the AVD offscreen"
Original message:
Though the View is on screen, the AVD itself has a matrix. Verify that
we respect that matrix.

Custom MyProgressBar just does a translate that happens to be the width
of the screen of my test device.

Addendum:
The plan is to make the View larger in order so that I can translate and
still have it potentially within the View.