Make it possible to scroll the AVD onto the screen
Place TextView into a HorizontalScrollView and make it exactly as wide
as my Pixel 3 screen. Put the ProgressBar to the right of the TextView,
so it is off screen. As hoped, even with the moved call to
setPropertyChangeWillBeConsumed, it still animates when I pull it back
onto the screen.

Now the thing to check is what happens if the AVD is offscreen, but the
View is on screen.